Authorization

+++ app/Http/Controllers/TaskController.php
@@ -23,6 +23,8 @@ class TaskController extends Controller
      */
     public function create()
     {
+        $this->authorize('manage tasks');
+
         return view('tasks.create');
     }
 
@@ -31,6 +33,8 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        $this->authorize('manage tasks');
+
         Task::create($request->validated());
 
         return redirect()->route('tasks.index');
@@ -49,6 +53,8 @@ class TaskController extends Controller
      */
     public function edit(Task $task)
     {
+        $this->authorize('manage tasks');
+
         return view('tasks.edit', compact('task'));
 
     }
@@ -58,6 +64,8 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        $this->authorize('manage tasks');
+
         $task->update($request->validated());
 
         return redirect()->route('tasks.index');
@@ -68,6 +76,7 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        $this->authorize('manage tasks');
         $task->delete();
 
         return redirect()->route('tasks.index');
